+++ test/cognito_user_exceptions_test.dart
@@ -98,6 +98,27 @@ void main() {
           e.toString(), equals('CognitoUserException: "SOFTWARE_TOKEN_MFA"'));
     }
   });
+  test('throw CognitoUserEmailOtpRequiredException generates message', () {
+    t() => throw CognitoUserEmailOtpRequiredException();
+    try {
+      t();
+    } on CognitoUserEmailOtpRequiredException catch (e) {
+      expect(
+          e.toString(), equals('CognitoUserException: "EMAIL_OTP"'));
+    }
+    try {
+      t();
+    } on CognitoUserException catch (e) {
+      expect(
+          e.toString(), equals('CognitoUserException: "EMAIL_OTP"'));
+    }
+    try {
+      t();
+    } catch (e) {
+      expect(
+          e.toString(), equals('CognitoUserException: "EMAIL_OTP"'));
+    }
+  });
   test('throw CognitoUserCustomChallengeException generates message', () {
     t() => throw CognitoUserCustomChallengeException();
     try {
@@ -135,6 +156,35 @@ void main() {
     } catch (e) {
       expect(e.toString(),
           equals('CognitoUserException: User Confirmation Necessary'));
+    }
+  });
+  test(
+      'throw CognitoUserPhoneNumberVerificationNecessaryException generates message',
+      () {
+    t() => throw CognitoUserPhoneNumberVerificationNecessaryException();
+    try {
+      t();
+    } on CognitoUserPhoneNumberVerificationNecessaryException catch (e) {
+      expect(
+          e.toString(),
+          equals(
+              'CognitoUserException: Verification of Attribute \'phone_number\' Necessary'));
+    }
+    try {
+      t();
+    } on CognitoUserException catch (e) {
+      expect(
+          e.toString(),
+          equals(
+              'CognitoUserException: Verification of Attribute \'phone_number\' Necessary'));
+    }
+    try {
+      t();
+    } catch (e) {
+      expect(
+          e.toString(),
+          equals(
+              'CognitoUserException: Verification of Attribute \'phone_number\' Necessary'));
     }
   });
 }